+++ chia-bls/Cargo.toml
@@ -9,11 +9,10 @@ homepage = "https://github.com/Chia-Network/chia_rs/chia-bls/"
 repository = "https://github.com/Chia-Network/chia_rs/chia-bls/"
 
 [features]
-py-bindings = ["dep:pyo3", "chia_py_streamable_macro", "chia-traits/py-bindings", "clvm-traits/py-bindings"]
+py-bindings = ["dep:pyo3", "chia_py_streamable_macro", "chia-traits/py-bindings"]
 
 [dependencies]
 chia-traits = { version = "=0.3.0", path = "../chia-traits" }
-clvm-traits = { version = "=0.2.14", features = ["derive"] }
 chia_py_streamable_macro = { version = "=0.3.0", path = "../chia_py_streamable_macro", optional = true }
 tiny-bip39 = "1.0.0"
 anyhow = "1.0.71"
@@ -29,7 +28,6 @@ arbitrary = { version = "1.3.0" }
 rand = "0.8.5"
 criterion = "0.5.1"
 rstest = "0.17.0"
-clvmr = "=0.3.0"
 
 [lib]
 crate-type = ["rlib"]
