+++ chia-bls/src/signature.rs
@@ -1,7 +1,6 @@
 use crate::{Error, GTElement, PublicKey, Result, SecretKey};
 use blst::*;
 use chia_traits::{read_bytes, Streamable};
-use clvm_traits::{ClvmDecoder, ClvmEncoder, FromClvm, FromClvmError, ToClvm, ToClvmError};
 use sha2::{Digest, Sha256};
 use std::borrow::Borrow;
 use std::convert::AsRef;
@@ -252,30 +251,6 @@ impl FromJsonDict for Signature {
                 .try_into()
                 .unwrap(),
         )?)
-    }
-}
-
-impl<N> FromClvm<N> for Signature {
-    fn from_clvm(
-        decoder: &impl ClvmDecoder<Node = N>,
-        node: N,
-    ) -> std::result::Result<Self, FromClvmError> {
-        let bytes = decoder.decode_atom(&node)?;
-        let error = Err(FromClvmError::WrongAtomLength {
-            expected: 96,
-            found: bytes.len(),
-        });
-        let bytes = bytes.try_into().or(error)?;
-        Self::from_bytes(bytes).map_err(|error| FromClvmError::Custom(error.to_string()))
-    }
-}
-
-impl<N> ToClvm<N> for Signature {
-    fn to_clvm(
-        &self,
-        encoder: &mut impl ClvmEncoder<Node = N>,
-    ) -> std::result::Result<N, ToClvmError> {
-        encoder.encode_atom(&self.to_bytes())
     }
 }
 
@@ -531,7 +506,6 @@ pub fn sign<Msg: AsRef<[u8]>>(sk: &SecretKey, msg: Msg) -> Signature {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use clvmr::Allocator;
     use hex::FromHex;
     use rand::rngs::StdRng;
     use rand::{Rng, SeedableRng};
@@ -1061,29 +1035,6 @@ mod tests {
         assert_eq!(
             format!("{:?}", sig),
             format!("<G2Element {}>", hex::encode(data))
-        );
-    }
-
-    #[test]
-    fn test_to_from_clvm() {
-        let mut a = Allocator::new();
-        let bytes = hex::decode("b45825c0ee7759945c0189b4c38b7e54231ebadc83a851bec3bb7cf954a124ae0cc8e8e5146558332ea152f63bf8846e04826185ef60e817f271f8d500126561319203f9acb95809ed20c193757233454be1562a5870570941a84605bd2c9c9a").expect("hex::decode()");
-        let ptr = a.new_atom(&bytes).expect("new_atom");
-
-        let sig = Signature::from_clvm(&a, ptr).expect("from_clvm");
-        assert_eq!(&sig.to_bytes()[..], &bytes[..]);
-
-        let sig_ptr = sig.to_clvm(&mut a).expect("to_clvm");
-        assert!(a.atom_eq(sig_ptr, ptr));
-    }
-
-    #[test]
-    fn test_from_clvm_failure() {
-        let mut a = Allocator::new();
-        let ptr = a.new_pair(a.one(), a.one()).expect("new_pair");
-        assert_eq!(
-            Signature::from_clvm(&a, ptr).unwrap_err(),
-            FromClvmError::ExpectedAtom
         );
     }
